+++ HISTORY.md
@@ -6,6 +6,9 @@ No advantage over using file_get_contents() and parsing contents.
 
 Instead using PHP gererator function to split processing to discrete chunks so that only the current row is in memory.
 
+# 2019-12-30 Check date fields
+- add test for date columns created, updated, let-date-available
+
 # 2019-12-28 Add testing for enums
 - Add test for enums such as STATUS_ID